fix(forms): GhostInput clipping/flicker, React 19 ref, own folder
Auto-size off a measured span (useLayoutEffect, +2px for the caret) to fix
text clipping and fast-typing flicker; drop forwardRef for the React 19 ref
prop; move into its own folder with a barrel.
