+++ frontend/documentation/components/GhostInput.stories.tsx
@@ -3,13 +3,14 @@ import type { Meta, StoryObj } from 'storybook'
 
 import GhostInput from 'components/base/forms/GhostInput'
 
-const meta: Meta = {
+const meta: Meta<typeof GhostInput> = {
+  component: GhostInput,
   parameters: { layout: 'centered' },
   title: 'Components/Forms/GhostInput',
 }
 export default meta
 
-type Story = StoryObj
+type Story = StoryObj<typeof GhostInput>
 
 const Interactive = () => {
   const [value, setValue] = useState('my-feature-flag')
@@ -29,5 +30,16 @@ export const Default: Story = {
 export const Empty: Story = {
   render: () => (
     <GhostInput value='' onChange={() => {}} placeholder='Enter a name...' />
+  ),
+}
+
+// Guards the clipping regression: the whole value must render, not "show_demo_butto".
+export const LongValue: Story = {
+  render: () => (
+    <GhostInput
+      value='show_demo_button'
+      onChange={() => {}}
+      placeholder='Enter a name...'
+    />
   ),
 }
