+++ frontend/web/components/base/forms/GhostInput/index.ts
@@ -1,0 +1,1 @@
+export { default } from './GhostInput'
